+++ public/front-roms-library/src/RomsLibrary.tsx
@@ -1,16 +1,16 @@
 // import { AddRom } from './components/AddRom/AddRom';
 import { RomsGrid, LoadingSpinner, RomsFilter } from "./components";
-import { useFetchRoms } from "./hooks";
+import { RomsPagination } from "./components/RomsPagination/RomsPagination";
+import { useFetchRoms, useFilterRoms } from "./hooks";
 
 
 export const RomsLibrary = () => {
+  const {videoconsoles} = useFilterRoms();
   const {roms, getNewRoms, isLoading} = useFetchRoms();
-
   // const onNewRom = (newRom: Rom) => {
   //   if (roms.find((rom) => newRom.title.toLowerCase() === rom.title.toLowerCase())) return;
   //   setRoms([...roms, newRom]);
   // }
-
 
   return (
     <>
@@ -24,6 +24,8 @@ export const RomsLibrary = () => {
       }
       
       <RomsGrid roms={roms} />
+
+      <RomsPagination getNewRoms={getNewRoms}/>
 
     </>
   );