+++ public/front-roms-library/src/components/RomsPagination/RomsPagination.tsx
@@ -1,0 +1,11 @@
+import { GetNewRoms } from "../../types"
+
+interface RomsPaginationProps {
+    getNewRoms: GetNewRoms,
+}
+
+export const RomsPagination: React.FC<RomsPaginationProps> = ({getNewRoms}) => {
+  return (
+    <div>RomsPagination</div>
+  )
+}