+++ public/front-roms-library/src/hooks/useFetchRoms.ts
@@ -1,12 +1,16 @@
 import { useEffect, useState } from "react";
 import { getRoms } from "../helpers";
 import { Rom, getRomsArgs } from "../types";
+import { useFilterRoms } from "./useFilterRoms";
 
 export const useFetchRoms = () => {
     const [roms, setRoms] = useState<Rom[]>([]);
     const [isLoading, setIsLoading] = useState<boolean>(true);
+    const {videoconsoles} = useFilterRoms();
 
-    const getNewRoms = async(getRomsArgs: getRomsArgs = {}) => {
+    const getNewRoms = async(getRomsArgs: getRomsArgs = {
+        videoconsoles
+    }) => {
         const newRoms = await getRoms(getRomsArgs);
         setRoms(newRoms);
         setIsLoading(false);
@@ -14,7 +18,7 @@ export const useFetchRoms = () => {
   
     useEffect(() => {
         getNewRoms();
-    }, []);
+    }, [videoconsoles]);
     
     return{
         // roms: roms, <--- This is the same than below